refactor(html): cleanup definition of tags
