+++ services/viva-ms/src/types/caseItem.ts
@@ -136,68 +136,3 @@ export interface CaseFormEncryption {
   type: EncryptionType.Decrypted;
   symmetricKeyName?: string;
 }
-
-// TOOD: probably should not be here
-export type ValidTags =
-  | 'address'
-  | 'aid'
-  | 'akassa'
-  | 'amount'
-  | 'annat'
-  | 'applicant'
-  | 'assets'
-  | 'boende'
-  | 'childcare'
-  | 'children'
-  | 'citizenship'
-  | 'coapplicant'
-  | 'csn'
-  | 'date'
-  | 'debtRent'
-  | 'description'
-  | 'electricity'
-  | 'email'
-  | 'expenses'
-  | 'fastighet'
-  | 'firstName'
-  | 'fordon'
-  | 'foreignPension'
-  | 'fulltime'
-  | 'homeinsurance'
-  | 'homelessDescription'
-  | 'housing'
-  | 'incomes'
-  | 'insurance'
-  | 'internet'
-  | 'lastName'
-  | 'layout'
-  | 'lon'
-  | 'medicin'
-  | 'message'
-  | 'note'
-  | 'notes'
-  | 'numberPeopleLiving'
-  | 'occupation'
-  | 'other'
-  | 'otherLivingDescription'
-  | 'otheroccupation'
-  | 'övrig'
-  | 'ownerContractApproved'
-  | 'ownRoom'
-  | 'parentalleave'
-  | 'parttime'
-  | 'pension'
-  | 'personalNumber'
-  | 'phonenumber'
-  | 'postalAddress'
-  | 'postalCode'
-  | 'rent'
-  | 'resident'
-  | 'school'
-  | 'sickleave'
-  | 'studies'
-  | 'tandvard'
-  | 'type'
-  | 'unemployed'
-  | 'unemployment'
-  | 'value';
